Rename platform -> platforms to fix tests/imports.
* When bazel runs tests (e.g. db_test.py) the directory tensorboard/
  gets added to the Python path.
* So if tensorboard/platform/__init__.py exists then
  import platform will end up importing tensorboard/platform and shadowing
  the system platform.
* We want that init file because we want to be able to run the code
  just by adding tensorboard to the path to facilitate debugging in IDEs.
* Simple fix is to just rename platform->platforms to avoid this conflict.
